(#178) Separate shell.cpp into 2 files
